+++ analyzer/AmetrinOptionalAnalyzer.cs
@@ -1,0 +1,71 @@
+﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
+using Microsoft.CodeAnalysis.Operations;
+using static Ametrin.Optional.Analyzer.Helper;
+
+namespace Ametrin.Optional.Analyzer;
+
+[DiagnosticAnalyzer(LanguageNames.CSharp)]
+public sealed class AmetrinOptionalAnalyzer : DiagnosticAnalyzer
+{
+
+    public static readonly DiagnosticDescriptor ImpossibleRequire
+        = new(id: "AmOptional001", title: "Impossible Require call", messageFormat: "{0} can never be {1}. If a conversion exists use .Map to explicitly use it.", category: "Usage", defaultSeverity: DiagnosticSeverity.Error, isEnabledByDefault: true);
+
+    public static readonly DiagnosticDescriptor UnnecessaryRequire
+        = new(id: "AmOptional002", title: "Unnecessary Require call", messageFormat: "type already is {0}", category: "Usage", defaultSeverity: DiagnosticSeverity.Warning, isEnabledByDefault: true);
+
+    public static readonly DiagnosticDescriptor WrongConditionalType
+        = new(id: "AmOptional003", title: "Wrong conditional return type", messageFormat: "default means {0} instead of {1}", category: "Usage", defaultSeverity: DiagnosticSeverity.Info, isEnabledByDefault: true);
+
+    public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; } = [ImpossibleRequire, UnnecessaryRequire, WrongConditionalType];
+
+    public override void Initialize(AnalysisContext context)
+    {
+        context.EnableConcurrentExecution();
+        context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.Analyze | GeneratedCodeAnalysisFlags.ReportDiagnostics);
+
+
+        context.RegisterOperationAction(static context =>
+        {
+            var invocation = (IInvocationOperation)context.Operation;
+            var targetMethod = invocation.TargetMethod;
+
+            if (IsRequireCastMethod(targetMethod))
+            {
+                var currentType = targetMethod.ContainingType.TypeArguments[0];
+                var targetType = targetMethod.TypeArguments[0];
+
+                var conversion = context.Compilation.ClassifyConversion(currentType, targetType);
+                if (conversion.IsIdentity)
+                {
+                    context.ReportDiagnostic(Diagnostic.Create(UnnecessaryRequire, invocation.Syntax.GetLocation(), targetType.ToDisplayString()));
+                }
+                else if (!conversion.Exists || conversion.IsUserDefined)
+                {
+                    context.ReportDiagnostic(Diagnostic.Create(ImpossibleRequire, invocation.Syntax.GetLocation(), currentType.ToDisplayString(), targetType.ToDisplayString()));
+                }
+            }
+
+        }, OperationKind.Invocation);
+
+        context.RegisterOperationAction(static context =>
+        {
+            var conversion = (IConversionOperation)context.Operation;
+
+            if (!conversion.IsImplicit) return;
+
+            if (conversion.Type is not INamedTypeSymbol target || !IsOptionalType(target)) return;
+
+            if (conversion.Operand is not IConditionalOperation condition || IsOptionalType(condition.Type!)) return;
+
+            var hasDefault = IsDefaultLiteral(condition.WhenTrue) || IsDefaultLiteral(condition.WhenFalse!);
+            if (!hasDefault) return;
+
+            context.ReportDiagnostic(Diagnostic.Create(WrongConditionalType, condition.Syntax.GetLocation(), condition.Type!.ToDisplayString(), conversion.Type.ToDisplayString()));
+
+        }, OperationKind.Conversion);
+    }
+}